refactor: add the CliLoop struct
