+++ applications/tari_base_node/src/commands/cli_loop.rs
@@ -1,0 +1,216 @@
+use std::time::{Duration, Instant};
+
+use crossterm::{
+    cursor,
+    event::{Event, EventStream, KeyCode, KeyEvent, KeyModifiers},
+    terminal,
+};
+use futures::{FutureExt, StreamExt};
+use rustyline::{config::OutputStreamType, error::ReadlineError, CompletionType, Config, EditMode, Editor};
+use tari_shutdown::ShutdownSignal;
+use tokio::{signal, time};
+
+use crate::{
+    commands::{
+        cli,
+        command::{CommandContext, WatchCommand},
+        parser::Parser,
+        reader::CommandReader,
+    },
+    LOG_TARGET,
+};
+
+pub struct CliLoop {
+    context: CommandContext,
+    reader: CommandReader,
+    commands: Vec<String>,
+    watch_task: Option<WatchCommand>,
+    non_interactive: bool,
+    first_signal: bool,
+    done: bool,
+    shutdown_signal: ShutdownSignal,
+}
+
+impl CliLoop {
+    pub fn new(context: CommandContext, watch_command: Option<String>, non_interactive: bool) -> Self {
+        let parser = Parser::new();
+        let commands = parser.get_commands();
+        let cli_config = Config::builder()
+            .history_ignore_space(true)
+            .completion_type(CompletionType::List)
+            .edit_mode(EditMode::Emacs)
+            .output_stream(OutputStreamType::Stdout)
+            .auto_add_history(true)
+            .build();
+        let mut rustyline = Editor::with_config(cli_config);
+        rustyline.set_helper(Some(parser));
+        let reader = CommandReader::new(rustyline);
+        let watch_task = {
+            if let Some(line) = watch_command {
+                WatchCommand::new(line)
+            } else if non_interactive {
+                WatchCommand::new("status --output log")
+            } else {
+                WatchCommand::new("status")
+            }
+        };
+        let shutdown_signal = context.shutdown.to_signal();
+        Self {
+            context,
+            reader,
+            commands,
+            watch_task: Some(watch_task),
+            non_interactive,
+            first_signal: false,
+            done: false,
+            shutdown_signal,
+        }
+    }
+
+    async fn run_watch_task(&mut self) {
+        if let Some(command) = self.watch_task.take() {
+            let start_time = Instant::now();
+            let mut interrupt = signal::ctrl_c().fuse().boxed();
+            let mut software_update_notif = self.context.software_updater.new_update_notifier().clone();
+            let config = self.context.config.clone();
+            let line = command.line();
+            let interval = command
+                .interval
+                .map(Duration::from_secs)
+                .unwrap_or(config.base_node_status_line_interval);
+            if let Err(err) = self.context.handle_command_str(line).await {
+                println!("Wrong command to watch `{}`. Failed with: {}", line, err);
+            } else {
+                let mut events = EventStream::new();
+                loop {
+                    terminal::enable_raw_mode().ok();
+                    let interval = get_status_interval(start_time, interval);
+                    tokio::select! {
+                        _ = interval => {
+                            terminal::disable_raw_mode().ok();
+                            if let Err(err) = self.context.handle_command_str(line).await {
+                                println!("Watched command `{}` failed: {}", line, err);
+                            }
+                            continue;
+                        },
+                        _ = &mut interrupt => {
+                            break;
+                        }
+                        event = events.next() => {
+                            match event {
+                                Some(Ok(Event::Key(key))) => {
+                                    match key {
+                                        KeyEvent { code: KeyCode::Char('c'), modifiers: KeyModifiers::CONTROL } => {
+                                            break;
+                                        }
+                                        _ => {
+                                            if self.non_interactive {
+                                                println!("Press Ctrl-C to interrupt the node.");
+                                            } else {
+                                                println!("Press Ctrl-C to enter the interactive shell.");
+                                            }
+                                        }
+                                    }
+                                }
+                                _ => {
+                                }
+                            }
+                        }
+                        // TODO: Is that good idea? Or add a separate command?
+                        Ok(_) = software_update_notif.changed() => {
+                            if let Some(ref update) = *software_update_notif.borrow() {
+                                println!(
+                                    "Version {} of the {} is available: {} (sha: {})",
+                                    update.version(),
+                                    update.app(),
+                                    update.download_url(),
+                                    update.to_hash_hex()
+                                );
+                            }
+                        }
+                    }
+                    crossterm::execute!(std::io::stdout(), cursor::MoveToNextLine(1)).ok();
+                }
+                terminal::disable_raw_mode().ok();
+            }
+        }
+    }
+
+    async fn execute_command(&mut self) {
+        tokio::select! {
+            res = self.reader.next_command() => {
+                if let Some(event) = res {
+                    match event {
+                        Ok(line) => {
+                            self.first_signal = false;
+                            if !line.is_empty() {
+                                match self.context.handle_command_str(&line).await {
+                                    Err(err) => {
+                                        println!("Command `{}` failed: {}", line, err);
+                                    }
+                                    Ok(command) => {
+                                        self.watch_task = command;
+                                    }
+                                }
+                            }
+                        }
+                        Err(ReadlineError::Interrupted) => {
+                            // If `Ctrl-C` is pressed
+                            if !self.first_signal {
+                                println!("Are you leaving already? Press Ctrl-C again (or Ctrl-D) to terminate the node.");
+                                self.first_signal = true;
+                            } else {
+                                self.done = true;
+                            }
+                        }
+                        Err(ReadlineError::Eof) => {
+                            // If `Ctrl-D` is pressed
+                            self.done = true;
+                        }
+                        Err(err) => {
+                            // TODO: Not sure we have to break here
+                            // This happens when the node is shutting down.
+                            log::debug!(target:  LOG_TARGET, "Could not read line from rustyline:{}", err);
+                            self.done = true;
+                        }
+                    }
+                } else {
+                    self.done = true;
+                }
+            },
+            _ = self.shutdown_signal.wait() => {
+                self.done = true;
+            }
+        }
+    }
+}
+
+impl CliLoop {
+    /// Runs the Base Node CLI loop
+    /// ## Parameters
+    /// `parser` - The parser to process input commands
+    /// `shutdown` - The trigger for shutting down
+    ///
+    /// ## Returns
+    /// Doesn't return anything
+    pub async fn cli_loop(mut self) {
+        cli::print_banner(self.commands.clone(), 3);
+
+        // TODO: Check for a new version here
+        while !self.done {
+            self.run_watch_task().await;
+            if self.non_interactive {
+                break;
+            }
+            self.execute_command().await;
+        }
+    }
+}
+
+fn get_status_interval(start_time: Instant, long_interval: Duration) -> time::Sleep {
+    let duration = match start_time.elapsed().as_secs() {
+        0..=120 => Duration::from_secs(5),
+        _ => long_interval,
+    };
+    time::sleep(duration)
+}
